requirements.txt
numpy>=1.26
keyboard>=0.13.5
mss>=6.1
PySimpleGUI>=4.60
Pillow>=10.0

